Fix crucial 64-bit correctness errors that would have caused catastrophic failure when a simulation has more than 4 billion players.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@9851 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_chart.cpp
@@ -192,7 +192,7 @@ int chart_t::raid_aps( std::vector<std::string>& images,
                        std::vector<player_t*> players_by_aps,
                        bool dps )
 {
-  int num_players = ( int ) players_by_aps.size();
+  int num_players = static_cast<int>( players_by_aps.size() );
 
   if ( num_players == 0 )
     return 0;
@@ -220,7 +220,7 @@ int chart_t::raid_aps( std::vector<std::string>& images,
       break;
     }
   }
-  num_players = player_list.size();
+  num_players = static_cast<int>( player_list.size() );
 
   if ( num_players == 0 )
     return 0;
